+++ notifications/CanceledEvent.php
@@ -48,7 +48,7 @@ class CanceledEvent extends BaseNotification
         if($this->source->content->container instanceof Space) {
             return Yii::t('CalendarModule.notifications_views_CanceledEvent', '{displayName} canceled event "{contentTitle}" in space {spaceName}.', [
                 'displayName' => Html::tag('strong', Html::encode($this->originator->displayName)),
-                'contentTitle' => Html::tag($this->getContentInfo($this->source, false)),
+                'contentTitle' => $this->getContentInfo($this->source, false),
                 'spaceName' =>  Html::encode($this->source->content->container->displayName)
             ]);
         } else {